Create consistently-named modules
The goal of this is to make `srb rbi gems` (from the 'sorbet' gem) is
repeatable when runs include polyfill (via sorbet-rails).

currently because the name is based on the object_id and the whims of
the universe this isn't the case.

The added benefit is we can reuse modules defined with the same
arguments. (this might be the primary reason to do this, but for me it's
a nice bonus). Also this will no doubt affect whatever tooling is
available for rbs files in ruby 3.0.0

because these methods are created arbitrarily based on arguments
it seems like the safest thing to name these is a base64 encoded version
of those arguments. (or close to base 64, +-/= are invalid name
characters, so i change the invalid characters to 1 and 2 (preceeded
by _ because we're not using it otherwise, and it means that no matter
the input, it will be only created by an identical string)

---

to show the objects it generates are repeatable across ruby runs

1. get sorbet gem
2. run `srb rbi hidden-definitions`
3. commit the generated files
4. run `srb rbi hidden-definitions` again.
5. there should be no change to sorbet/rbi/hidden-definitions/hidden.rbi